StudentManage

+++ src/main/java/com/StuMng/po/StuMarkInfo.java
@@ -1,6 +1,5 @@
 package com.StuMng.po;
 
-import java.math.BigDecimal;
 import java.util.Date;
 
 public class StuMarkInfo {
@@ -10,11 +9,11 @@ public class StuMarkInfo {
 
     private String coursename;
 
-    private BigDecimal basescore;
+    private Double basescore;
 
-    private BigDecimal testscore;
+    private Double testscore;
 
-    private BigDecimal finalscore;
+    private Double finalscore;
 
     private Date adddate;
 
@@ -44,27 +43,27 @@ public class StuMarkInfo {
         this.coursename = coursename == null ? null : coursename.trim();
     }
 
-    public BigDecimal getBasescore() {
+    public Double getBasescore() {
         return basescore;
     }
 
-    public void setBasescore(BigDecimal basescore) {
+    public void setBasescore(Double basescore) {
         this.basescore = basescore;
     }
 
-    public BigDecimal getTestscore() {
+    public Double getTestscore() {
         return testscore;
     }
 
-    public void setTestscore(BigDecimal testscore) {
+    public void setTestscore(Double testscore) {
         this.testscore = testscore;
     }
 
-    public BigDecimal getFinalscore() {
+    public Double getFinalscore() {
         return finalscore;
     }
 
-    public void setFinalscore(BigDecimal finalscore) {
+    public void setFinalscore(Double finalscore) {
         this.finalscore = finalscore;
     }
